Move neat_resolver to correct header
struct neat_resolver is now private to neat_resolver.c so move it to
neat_resolver.h
