+++ neat_he.c
@@ -224,9 +224,6 @@ neat_error_code neat_he_lookup(neat_ctx *ctx, neat_flow *flow, uv_poll_cb callba
         ctx->resolver = neat_resolver_init(ctx, "/etc/resolv.conf");
     }
 
-    /*ctx->resolver->userData1 = (void *)flow; // TODO: This doesn't allow multiple sockets
-    ctx->resolver->userData2 = callback_fx;*/
-
     /* FIXME: derivation of the socket type is wrong.
      * FIXME: Make use of the array of protocols
      */
